+++ src/imio/smartweb/core/webcomponents/src/components/News/News.jsx
@@ -75,7 +75,7 @@ const NewsView = (props) => {
 
     // set batch
     const loadMore = () => {
-        setBatchStart((batchStart) => batchStart + props.batchSize);
+        setBatchStart((batchStart) => batchStart + parseInt(props.batchSize));
         setLoadMoreLaunch(true);
     };
     // Update filters Batch